Add files via upload
In this project, project build string concept, the program prompts the user to enter a text. It then uses various methods to analyze the text and provides information such as the word count, character count, and average word length.
The countWords method splits the text into words using regular expressions and returns the number of words. The countCharacters method simply returns the length of the text. The calculateAverageWordLength method calculates the average word length by dividing the total character count by the word count.